Split landmarks tabs into server panels and fix tab focus ring
LandmarksTabs was one big client component wrapping every landmark
image and eatery card in client JS just to get tab switching. Split
it into a thin client TabsShell plus three server-rendered panels
(visit, eat, around) passed in as children.

Also fixed .nn-tab silently killing the focus-visible ring: box-shadow
was force-disabled at all times, so keyboard focus had no indicator on
either the landmarks or check-in tabs. Replaced with a visible outline.

Landmark.image/mapsUrl and Eatery.mapsUrl are required again since
every entry sets them; the optional-with-fallback shape had no real
use case yet.
